Edit and delete authorizated

+++ smart-real-estate-management-system/src/app/components/estate-detail/estate-detail.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { EstateService } from '../../services/estate.service';
 import { CommonModule } from '@angular/common';
-
+import { UserService } from '../../services/user.service';
 
 @Component({
   selector: 'app-estate-detail',
@@ -17,7 +17,8 @@ export class EstateDetailComponent implements OnInit {
   constructor(
     private readonly route: ActivatedRoute,
     private readonly estateService: EstateService,
-    private readonly router: Router
+    private readonly router: Router,
+    private readonly userService: UserService
   ) {}
 
   ngOnInit(): void {
@@ -30,6 +31,13 @@ export class EstateDetailComponent implements OnInit {
   }
 
   onDelete(id: string): void {
+    const userId=this.userService.getUserId();
+
+    if (userId!=this.estate.userId){
+      console.error('This user is unauthorized to delete this estate.');
+      return;
+    }
+
     this.estateService.deleteEstate(id).subscribe(() => {
       this.router.navigate(['/estates/paginated']);
     });
